build(deps): bump EDC to 0.18.0
